db/requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
numpy==2.1.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2